# apps/api/openapi.json
{
 "openapi": "3.0.3",
 "info": {
  "title": "Intuitive Fusion Procurement Portal API (POC)",
  "version": "0.1.0-draft",
  "description": "REST API for the POC. Auth is mock (HttpOnly session cookie) behind an IdentityProvider interface. All mutations emit audit events. Errors follow RFC 7807 (application/problem+json) with stable `code` and `correlationId`."
 },
 "servers": [
  {
   "url": "http://localhost:4000/api/v1",
   "description": "Local dev"
  }
 ],
...
 "security": [
  {
   "sessionCookie": []
  }
 ],
 "paths": {
...
  "/contracts/{id}/milestones": {
   "put": {
    "operationId": "setContractMilestones",
    "tags": [
     "Contracts"
    ],
    "summary": "Replace the milestones of an executed contract (within its term); scheduled reminders follow",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid"
      }
     }
    ],
    "x-roles": [
     "CONTRACT_MGR",
     "LEGAL",
     "PROCUREMENT"
    ],
    "responses": {
     "200": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Contract"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "403": {
      "$ref": "#/components/responses/E403"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     },
     "409": {
      "$ref": "#/components/responses/E409"
     },
     "422": {
      "$ref": "#/components/responses/E422"
     }
    },
    "description": "422 MILESTONE_OUTSIDE_TERM",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ContractMilestones"
       }
      }
     }
    }
   }
  },
  "/contracts/{id}/extensions": {
   "put": {
    "operationId": "setContractExtensions",
    "tags": [
     "Contracts"
    ],
    "summary": "Replace the optional extensions (months) of an executed contract",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid"
      }
     }
    ],
    "x-roles": [
     "CONTRACT_MGR",
     "LEGAL",
     "PROCUREMENT"
    ],
    "responses": {
     "200": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Contract"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "403": {
      "$ref": "#/components/responses/E403"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     },
     "409": {
      "$ref": "#/components/responses/E409"
     },
     "422": {
      "$ref": "#/components/responses/E422"
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ContractExtensions"
       }
      }
     }
    }
   }
  },
  "/contracts/{id}/owner": {
   "put": {
    "operationId": "setContractOwner",
    "tags": [
     "Contracts"
    ],
    "summary": "Change the contract owner to a contract manager",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid"
      }
     }
    ],
    "x-roles": [
     "CONTRACT_MGR",
     "LEGAL",
     "PROCUREMENT"
    ],
    "responses": {
     "200": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Contract"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "403": {
      "$ref": "#/components/responses/E403"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     },
     "409": {
      "$ref": "#/components/responses/E409"
     },
     "422": {
      "$ref": "#/components/responses/E422"
     }
    },
    "description": "422 NOT_A_CONTRACT_MANAGER",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ContractOwner"
       }
      }
     }
    }
   }
  },
  "/contracts/{id}/variations": {
   "post": {
    "operationId": "createVariation",
    "tags": [
     "Contracts"
    ],
    "summary": "Create a variation of an executed contract: a child contract with cumulative value tracking, drafted, reviewed and signed like any contract",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid"
      }
     }
    ],
    "x-roles": [
     "LEGAL",
     "PROCUREMENT"
    ],
    "responses": {
     "201": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Contract"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "403": {
      "$ref": "#/components/responses/E403"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     },
     "409": {
      "$ref": "#/components/responses/E409"
     },
     "422": {
      "$ref": "#/components/responses/E422"
     }
    },
    "description": "409 VARIATION_OPEN; 422 EMPTY_VARIATION; signing authority is judged on the cumulative value",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/VariationCreate"
       }
      }
     }
    }
   }
  },
  "/contracts/{id}/deviations/{clauseId}/risk": {
   "put": {
    "operationId": "setDeviationRisk",
    "tags": [
     "Contracts"
    ],
    "summary": "Legal amends the proposed risk rating of a deviation",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid"
      }
     },
     {
      "name": "clauseId",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "x-roles": [
     "LEGAL"
    ],
    "responses": {
     "200": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Contract"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "403": {
      "$ref": "#/components/responses/E403"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     },
     "409": {
      "$ref": "#/components/responses/E409"
     },
     "422": {
      "$ref": "#/components/responses/E422"
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DeviationRisk"
       }
      }
     }
    }
   }
  },
  "/contracts/{id}/deviations/{clauseId}/decision": {
   "post": {
    "operationId": "decideDeviation",
    "tags": [
     "Contracts"
    ],
    "summary": "A delegate approves or rejects a deviation; a mandatory or high-risk change must be approved before release",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid"
      }
     },
     {
      "name": "clauseId",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "x-roles": [
     "DELEGATE",
     "EXEC"
    ],
    "responses": {
     "200": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Contract"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "403": {
      "$ref": "#/components/responses/E403"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     },
     "409": {
      "$ref": "#/components/responses/E409"
     },
     "422": {
      "$ref": "#/components/responses/E422"
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/Decision"
       }
      }
     }
    }
   }
  },
...
  "/contracts/{id}/alerts": {
   "get": {
    "operationId": "listAlerts",
    "tags": [
     "Contracts"
    ],
    "summary": "Alerts of the contract with their delivery log (due alerts fire first)",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid"
      }
     }
    ],
    "x-roles": [
     "CONTRACT_MGR",
     "PROCUREMENT",
     "LEGAL",
     "EXEC"
    ],
    "responses": {
     "200": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/Alert"
         }
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "403": {
      "$ref": "#/components/responses/E403"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     }
    }
   },
   "post": {
    "operationId": "createAlert",
    "tags": [
     "Contracts"
    ],
    "summary": "Create a custom alert from a plain-language instruction such as: alert me 1 year before expiry and include whoever is my manager then. The recipients are resolved when it fires",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid"
      }
     }
    ],
    "x-roles": [
     "CONTRACT_MGR",
     "PROCUREMENT",
     "LEGAL",
     "EXEC"
    ],
    "responses": {
     "201": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AlertCreated"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "403": {
      "$ref": "#/components/responses/E403"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     },
     "409": {
      "$ref": "#/components/responses/E409"
     },
     "422": {
      "$ref": "#/components/responses/E422"
     }
    },
    "description": "422 ALERT_NOT_UNDERSTOOD explains what to type",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AlertCreate"
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "securitySchemes": {
   "sessionCookie": {
    "type": "apiKey",
    "in": "cookie",
    "name": "if_session"
   }
  },
  "schemas": {
...
   "Decision": {
    "type": "object",
    "properties": {
     "decision": {
      "type": "string",
      "enum": [
       "APPROVE",
       "REJECT"
      ]
     },
     "comment": {
      "type": "string",
      "maxLength": 2000
     },
     "gate": {
      "type": "string",
      "enum": [
       "RISK_SIGNOFF"
      ]
     }
    },
    "additionalProperties": false,
    "required": [
     "decision"
    ]
   },
...
   "ContractMilestones": {
    "type": "object",
    "properties": {
     "milestones": {
      "type": "array",
      "items": {
       "type": "object",
       "properties": {
        "title": {
         "type": "string"
        },
        "dueDate": {
         "type": "string",
         "format": "date"
        }
       },
       "additionalProperties": false,
       "required": [
        "title",
        "dueDate"
       ]
      }
     }
    },
    "additionalProperties": false,
    "required": [
     "milestones"
    ]
   },
   "ContractExtensions": {
    "type": "object",
    "properties": {
     "extensions": {
      "type": "array",
      "items": {
       "type": "integer",
       "minimum": 1,
       "maximum": 60
      }
     }
    },
    "additionalProperties": false,
    "required": [
     "extensions"
    ]
   },
   "ContractOwner": {
    "type": "object",
    "properties": {
     "ownerId": {
      "type": "string",
      "format": "uuid"
     }
    },
    "additionalProperties": false,
    "required": [
     "ownerId"
    ]
   },
   "VariationCreate": {
    "type": "object",
    "properties": {
     "reason": {
      "type": "string",
      "minLength": 10,
      "maxLength": 1000
     },
     "value": {
      "type": "number",
      "minimum": 0
     },
     "endDate": {
      "type": "string",
      "format": "date"
     }
    },
    "additionalProperties": false,
    "required": [
     "reason",
     "value"
    ]
   },
   "DeviationRisk": {
    "type": "object",
    "properties": {
     "risk": {
      "type": "string",
      "enum": [
       "LOW",
       "MEDIUM",
       "HIGH"
      ]
     }
    },
    "additionalProperties": false,
    "required": [
     "risk"
    ]
   },
...
   "Contract": {
    "type": "object",
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid"
     },
     "number": {
      "type": "string"
     },
     "tenderId": {
      "type": "string",
      "format": "uuid"
     },
     "supplierId": {
      "type": "string",
      "format": "uuid"
     },
     "templateId": {
      "type": "string"
     },
     "status": {
      "type": "string",
      "enum": [
       "DRAFT",
       "LEGAL_REVIEW",
       "AWAITING_SIGNATURE",
       "PARTIALLY_SIGNED",
       "EXECUTED"
      ]
     },
     "value": {
      "type": "number"
     },
     "startDate": {
      "type": "string",
      "format": "date"
     },
     "endDate": {
      "type": "string",
      "format": "date"
     },
     "noticeDays": {
      "type": "integer"
     },
     "clauses": {
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/Clause"
      }
     },
     "signatures": {
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/Approval"
      }
     },
     "parentId": {
      "type": "string",
      "format": "uuid"
     },
     "locked": {
      "type": "boolean"
     },
     "version": {
      "type": "integer"
     },
     "supplierName": {
      "type": "string"
     },
     "title": {
      "type": "string"
     },
     "requestNumber": {
      "type": "string"
     },
     "signed": {
      "type": "integer"
     },
     "signaturesRequired": {
      "type": "integer"
     },
     "deviations": {
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/ContractDeviation"
      }
     },
     "record": {
      "$ref": "#/components/schemas/ContractRecord"
     },
     "chain": {
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/ContractSigner"
      }
     },
     "permissions": {
      "$ref": "#/components/schemas/ContractPermissions"
     },
     "parent": {
      "type": "object",
      "properties": {
       "id": {
        "type": "string",
        "format": "uuid"
       },
       "number": {
        "type": "string"
       }
      },
      "additionalProperties": false
     },
     "variations": {
      "type": "array",
      "items": {
       "type": "object",
       "properties": {
        "id": {
         "type": "string",
         "format": "uuid"
        },
        "number": {
         "type": "string"
        },
        "status": {
         "type": "string"
        },
        "value": {
         "type": "number"
        },
        "endDate": {
         "type": "string",
         "format": "date"
        }
       },
       "additionalProperties": false
      }
     },
     "cumulative": {
      "type": "object",
      "properties": {
       "value": {
        "type": "number"
       },
       "endDate": {
        "type": "string",
        "format": "date"
       }
      },
      "additionalProperties": false
     },
     "deviationBlockers": {
      "type": "array",
      "items": {
       "type": "string"
      }
     }
    },
    "additionalProperties": false,
    "required": [
     "id",
     "number",
     "status"
    ]
   },
...
   "Alert": {
    "type": "object",
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid"
     },
     "contractId": {
      "type": "string",
      "format": "uuid"
     },
     "kind": {
      "type": "string",
      "enum": [
       "EXPIRY",
       "NOTICE",
       "MILESTONE",
       "EXTENSION",
       "CUSTOM"
      ]
     },
     "triggerDate": {
      "type": "string",
      "format": "date"
     },
     "recipientRule": {
      "type": "string"
     },
     "status": {
      "type": "string",
      "enum": [
       "SCHEDULED",
       "SENT",
       "CANCELLED"
      ]
     },
     "origin": {
      "type": "string",
      "enum": [
       "SYSTEM",
       "USER"
      ]
     },
     "sentAt": {
      "type": "string",
      "format": "date-time"
     },
     "contractNumber": {
      "type": "string"
     },
     "endDate": {
      "type": "string",
      "format": "date"
     },
     "deliveries": {
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/AlertDelivery"
      }
     }
    },
    "additionalProperties": false,
    "required": [
     "id",
     "contractId",
     "kind",
     "triggerDate"
    ]
   },
...
   "AlertCreate": {
    "type": "object",
    "properties": {
     "instruction": {
      "type": "string",
      "minLength": 5,
      "maxLength": 500
     }
    },
    "additionalProperties": false,
    "required": [
     "instruction"
    ]
   },
...
  },
  "responses": {
   "E400": {
    "description": "Validation error",
    "content": {
     "application/problem+json": {
      "schema": {
       "$ref": "#/components/schemas/Problem"
      }
     }
    }
   },
   "E401": {
    "description": "Not authenticated",
    "content": {
     "application/problem+json": {
      "schema": {
       "$ref": "#/components/schemas/Problem"
      }
     }
    }
   },
   "E403": {
    "description": "Forbidden",
    "content": {
     "application/problem+json": {
      "schema": {
       "$ref": "#/components/schemas/Problem"
      }
     }
    }
   },
   "E404": {
    "description": "Not found",
    "content": {
     "application/problem+json": {
      "schema": {
       "$ref": "#/components/schemas/Problem"
      }
     }
    }
   },
   "E409": {
    "description": "Conflict / rule violated",
    "content": {
     "application/problem+json": {
      "schema": {
       "$ref": "#/components/schemas/Problem"
      }
     }
    }
   },
   "E422": {
    "description": "Business rule failed",
    "content": {
     "application/problem+json": {
      "schema": {
       "$ref": "#/components/schemas/Problem"
      }
     }
    }
   },
...
  }
 }
}